Make the cache independent from the absolute path
Before this change, the cache was dependent on the full path to the service
(including the path the rails app was mounted, e.g. through
RAILS_RELATEIVE_URL_ROOT).

This made the cache unusable in case the mount point changed. This changes
makes the links from the change relative, fixing the issue with paths.

See https://bugzilla.redhat.com/show_bug.cgi?id=852388#c14 for example of such
an issue.